+++ packages/terra-date-input/tests/wdio/date-input-field-spec.js
@@ -116,7 +116,7 @@ Terra.describeViewports('Date Input Field', ['medium'], () => {
     });
 
     describe('Month Select Pressing DELETE in month with no value shifts focus to day in d-m-y format', () => {
-      it('should shift focus to day  on pressing delete', () => {
+      it('should shift focus to day on pressing delete', () => {
         browser.url('/raw/tests/terra-date-input/date-input/day-month-year-date-input-field');
         browser.refresh();
         // Removes the blinking cursor to prevent screenshot mismatches.
@@ -126,7 +126,7 @@ Terra.describeViewports('Date Input Field', ['medium'], () => {
         $('select[name="terra-date-month-date-input"]').click();
         browser.keys(['Enter']);
         browser.keys(['Delete']);
-        Terra.validates.element('shift focus to day');
+        Terra.validates.element('shift focus from month to day on delete');
       });
     });
 
@@ -173,7 +173,7 @@ Terra.describeViewports('Date Input Field', ['medium'], () => {
 
         $('input[name="terra-date-day-date-input"]').click();
         browser.keys(['ArrowUp']);
-        Terra.validates.element('increment day-empty');
+        Terra.validates.element('increment day empty');
       });
     });
 
@@ -188,7 +188,7 @@ Terra.describeViewports('Date Input Field', ['medium'], () => {
         $('input[name="terra-date-day-date-input"]').click();
         browser.keys(['31', 'ArrowUp']);
 
-        Terra.validates.element('cycle day');
+        Terra.validates.element('cycle up day');
       });
     });
 
@@ -203,7 +203,7 @@ Terra.describeViewports('Date Input Field', ['medium'], () => {
         $('input[name="terra-date-day-date-input"]').click();
         browser.keys(['ArrowDown']);
 
-        Terra.validates.element('decrement day-filled');
+        Terra.validates.element('decrement day filled');
       });
     });
 
@@ -218,7 +218,7 @@ Terra.describeViewports('Date Input Field', ['medium'], () => {
         $('input[name="terra-date-day-date-input"]').click();
         browser.keys(['ArrowDown']);
 
-        Terra.validates.element('decrement day-empty');
+        Terra.validates.element('decrement day empty');
       });
     });
 
@@ -232,7 +232,7 @@ Terra.describeViewports('Date Input Field', ['medium'], () => {
 
         $('input[name="terra-date-day-date-input"]').click();
         browser.keys(['01', 'ArrowDown']);
-        Terra.validates.element('cycle day');
+        Terra.validates.element('cycle down day');
       });
     });
 
@@ -294,7 +294,7 @@ Terra.describeViewports('Date Input Field', ['medium'], () => {
         $('input[name="terra-date-year-date-input"]').click();
         browser.keys(['ArrowUp']);
 
-        Terra.validates.element('increment year-filled');
+        Terra.validates.element('increment year filled');
       });
     });
 
@@ -309,7 +309,7 @@ Terra.describeViewports('Date Input Field', ['medium'], () => {
         $('input[name="terra-date-year-date-input"]').click();
         browser.keys(['ArrowUp']);
 
-        Terra.validates.element('increment year-empty');
+        Terra.validates.element('increment year empty');
       });
     });
 
@@ -324,7 +324,7 @@ Terra.describeViewports('Date Input Field', ['medium'], () => {
         $('input[name="terra-date-year-date-input"]').click();
         browser.keys(['2100', 'ArrowUp']);
 
-        Terra.validates.element('cycle year');
+        Terra.validates.element('cycle up year');
       });
     });
 
@@ -339,7 +339,7 @@ Terra.describeViewports('Date Input Field', ['medium'], () => {
         $('input[name="terra-date-year-date-input"]').click();
         browser.keys(['ArrowDown']);
 
-        Terra.validates.element('decrement year-filled');
+        Terra.validates.element('decrement year filled');
       });
     });
 
@@ -354,7 +354,7 @@ Terra.describeViewports('Date Input Field', ['medium'], () => {
         $('input[name="terra-date-year-date-input"]').click();
         browser.keys(['ArrowDown']);
 
-        Terra.validates.element('decrement year-empty');
+        Terra.validates.element('decrement year empty');
       });
     });
 
@@ -369,7 +369,7 @@ Terra.describeViewports('Date Input Field', ['medium'], () => {
         $('input[name="terra-date-year-date-input"]').click();
         browser.keys(['1900', 'ArrowDown']);
 
-        Terra.validates.element('cycle year');
+        Terra.validates.element('cycle down year');
       });
     });
 
@@ -399,7 +399,7 @@ Terra.describeViewports('Date Input Field', ['medium'], () => {
         $('input[name="terra-date-year-date-input"]').click();
         browser.keys(['Delete']);
 
-        Terra.validates.element('shift focus to day');
+        Terra.validates.element('shift focus from year to day on delete');
       });
     });