+++ packages/terra-abstract-modal/CHANGELOG.md
@@ -2,15 +2,17 @@
 
 ## Unreleased
 
+* Changed
+  * Updated to terra-functional-testing
+
 ## 3.31.0 - (January 5, 2021)
 
 * Changed
   * Updated to support a peer dependency of react-intl v2-v5
   * Updated FormattedMessage child function to join if given an array.
-  * Updated to terra-functional-testing
 
 * Fixed
- * Fixed broken links in documentation.
+  * Fixed broken links in documentation.
 
 ## 3.30.0 - (November 24, 2020)