+++ packages/terra-application-utility/CHANGELOG.md
@@ -12,7 +12,7 @@
   * Removed `intlShape` imports
 
 * Fixed
- * Fixed broken links in documentation.
+  * Fixed broken links in documentation.
 
 ## 2.40.0 - (November 24, 2020)